# frontend/lib/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "allwin API",
  "version": "1.0.0"
 },
 "paths": {
  "/api/v1/auth/methods": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Auth Methods",
    "operationId": "auth_methods_api_v1_auth_methods_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AuthMethodsDTO"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "410": {
      "description": "Gone",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "429": {
      "description": "Too Many Requests",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "502": {
      "description": "Bad Gateway",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/wechat/oa/start": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Wechat Oa Start",
    "description": "用户点击登录后进入(前端不得在页面加载时自动跳转到这里)。",
    "operationId": "wechat_oa_start_api_v1_auth_wechat_oa_start_get",
    "parameters": [
     {
      "name": "next",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "/",
       "title": "Next"
      }
     },
     {
      "name": "device",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Device"
      }
     }
    ],
    "responses": {
     "302": {
      "description": "跳转微信授权页(无 body)"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "410": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Gone"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "502": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Gateway"
     },
     "503": {
      "description": "微信登录暂未开放(AUTH_DISABLED)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/wechat/oa/callback": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Wechat Oa Callback",
    "operationId": "wechat_oa_callback_api_v1_auth_wechat_oa_callback_get",
    "parameters": [
     {
      "name": "code",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Code"
      }
     },
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "State"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WechatCallbackApprovedDTO"
        }
       }
      }
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "410": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Gone"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "502": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Gateway"
     },
     "302": {
      "description": "login 分支:种会话 Cookie 后跳回站内 next(无 body)"
     },
     "503": {
      "description": "微信登录暂未开放(AUTH_DISABLED)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/wechat/device": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Create Device Login",
    "operationId": "create_device_login_api_v1_auth_wechat_device_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceLoginCreatedDTO"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "410": {
      "description": "Gone",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "429": {
      "description": "Too Many Requests",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "502": {
      "description": "Bad Gateway",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "503": {
      "description": "微信登录暂未开放(AUTH_DISABLED)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/wechat/device/{request_id}/claim": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Claim Device Login",
    "operationId": "claim_device_login_api_v1_auth_wechat_device__request_id__claim_post",
    "parameters": [
     {
      "name": "request_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Request Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DeviceClaimBody"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceClaimResultDTO"
        }
       }
      }
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "410": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Gone"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     },
     "429": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Too Many Requests"
     },
     "502": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Gateway"
     },
     "503": {
      "description": "微信登录暂未开放(AUTH_DISABLED)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/password/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Password Login",
    "operationId": "password_login_api_v1_auth_password_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PasswordLoginBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OkDTO"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "410": {
      "description": "Gone",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "429": {
      "description": "Too Many Requests",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "502": {
      "description": "Bad Gateway",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/logout": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Logout",
    "operationId": "logout_api_v1_auth_logout_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OkDTO"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "410": {
      "description": "Gone",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "429": {
      "description": "Too Many Requests",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "502": {
      "description": "Bad Gateway",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_api_v1_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeDTO"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "410": {
      "description": "Gone",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "429": {
      "description": "Too Many Requests",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "502": {
      "description": "Bad Gateway",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ApiErrorDTO": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "details": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Details"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message",
     "details"
    ],
    "title": "ApiErrorDTO"
   },
...
   "AuthMethodsDTO": {
    "properties": {
     "wechat_enabled": {
      "type": "boolean",
      "title": "Wechat Enabled"
     }
    },
    "type": "object",
    "required": [
     "wechat_enabled"
    ],
    "title": "AuthMethodsDTO"
   },
...
   "DeviceClaimBody": {
    "properties": {
     "secret": {
      "type": "string",
      "title": "Secret"
     }
    },
    "type": "object",
    "required": [
     "secret"
    ],
    "title": "DeviceClaimBody"
   },
   "DeviceClaimResultDTO": {
    "properties": {
     "status": {
      "type": "string",
      "enum": [
       "pending",
       "claimed"
      ],
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "status"
    ],
    "title": "DeviceClaimResultDTO"
   },
   "DeviceLoginCreatedDTO": {
    "properties": {
     "request_id": {
      "type": "string",
      "title": "Request Id"
     },
     "secret": {
      "type": "string",
      "title": "Secret"
     },
     "qr_url": {
      "type": "string",
      "title": "Qr Url"
     },
     "expires_at": {
      "type": "string",
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "request_id",
     "secret",
     "qr_url",
     "expires_at"
    ],
    "title": "DeviceLoginCreatedDTO"
   },
...
   "MeDTO": {
    "properties": {
     "authenticated": {
      "type": "boolean",
      "title": "Authenticated"
     },
     "user": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/MeUser"
       },
       {
        "type": "null"
       }
      ]
     },
     "plan": {
      "type": "string",
      "title": "Plan"
     },
     "entitlements": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Entitlements"
     },
     "session_expires_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session Expires At"
     }
    },
    "type": "object",
    "required": [
     "authenticated",
     "plan",
     "entitlements"
    ],
    "title": "MeDTO"
   },
   "MeUser": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_name",
     "role"
    ],
    "title": "MeUser"
   },
...
   "OkDTO": {
    "properties": {
     "status": {
      "type": "string",
      "const": "ok",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "status"
    ],
    "title": "OkDTO"
   },
   "PasswordLoginBody": {
    "properties": {
     "username": {
      "type": "string",
      "title": "Username"
     },
     "password": {
      "type": "string",
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "PasswordLoginBody"
   },
...
   "WechatCallbackApprovedDTO": {
    "properties": {
     "status": {
      "type": "string",
      "const": "approved",
      "title": "Status"
     },
     "message": {
      "type": "string",
      "title": "Message"
     }
    },
    "type": "object",
    "required": [
     "status",
     "message"
    ],
    "title": "WechatCallbackApprovedDTO",
    "description": "oa/callback 的 device_approve 分支(login 分支为 302 跳转,无 body)。"
   },
...
  }
 }
}